changed bestsellerproducts as general comp
